requirements.txt
langchain==0.0.208
openai==0.27.0
python-dotenv==1.0.1
streamlit==1.31.1
azure-identity==1.12.0
azure-search-documents==11.4.0
fastapi==0.95.2
psycopg2-binary==2.9.6
sqlalchemy==1.4.47
tiktoken==0.4.0
uvicorn==0.22.0
tabulate==0.9.0
teradatasql==17.20.0.13
pandas==1.5.3
azure-keyvault-secrets==4.6.0
semantic-kernel==0.9.4b1
beautifulsoup4==4.12.2
azure-cognitiveservices-search-customsearch==2.0.0
llama-index==0.6.25
pyodbc==4.0.39
azure-functions==1.12.1